Add support for action GET_LANGUAGE_DETAILS
The nice thing is that Google Translate now displays the
microphone button next to the Estonian input box.

However, it does not display the microphone button for any other
languages anymore, even if Google Voice Search is set to be
the default. This looks like a bug in Google Translate (or
Google Voice Search).