Fix: server connection was not properly closed
Defensively checking that the audio byte array is non-empty meant
that occasionally the final (closing) chunk was not sent and thus
the server connection was not closed properly. The user could observe
this as non-stable behaviour where occasionally the transcription was not
returned for no apparent reason. (These occasions became more frequent once
we started pushing chunks with a smaller time interval.)