Make it work with Google Chrome
which does not seem to accept the results from the recognizer,
probably expecting it to perform the web search itself.
Chrome specifies the extras "WEB_SEARCH_ONLY" and "LANGUAGE_MODEL"="web_search",
but sets the action to "RECOGNIZE_SPEECH" (rather than "WEB_SEARCH"),
so maybe it's a bug in Chrome.