MicButton generates rewrite utterances if touched
i.e. one can map mic button swipes, double taps,
and long presses to commands. Depending on the context
the mapping can be different, e.g. right swipe can be
"send" in a chat app, "go the the next search result" in a text editor,
"go to the next cell" in a spreadsheet, etc.

During recognition the panel now responds to the same touches as when
idle (i.e. OnCursorTouchListener by default). Before it always implemented
OnSwipeTouchListener, which was possibly confusing, and misfired often.