Support requesting Do Not Disturb access
The respective permission is declared in the manifest
and a shortcut to the respective settings is in the Kõnele developer menu.

It seems that this would only be needed if Kõnele is used
while background audio is playing and the app that plays it
ignores the request to pause it.

Having access to Do Not Disturb setting allows us to mute the audio
stream (on N and higher). This has always seemed to work on the phone,
but crashes on TV if Do Not Disturb access was not granted.