Improve combo button
If more than 2 combos are selected then show them as individual buttons. The button label is the locale ID or the first 3 letters of the service name in case locale is "und" (undefined). Long-pressing on the button shows the full label.

Also, long-pressing on the IME action button shows the action label (often missing).

Long-press is implemented via the tooltip, which is available only in Android v8+.